+++ src/vfs/shared.ts
@@ -1,7 +1,7 @@
 // Utilities and shared data
 
 import type * as fs from 'node:fs';
-import type { V_Context } from '../context.js';
+import type { V_Context } from '../internal/contexts.js';
 import type { FileSystem } from '../internal/filesystem.js';
 import type { InodeLike } from '../internal/inode.js';
 import type { FSContext } from '../internal/contexts.js';
