+++ src/vfs/stats.ts
@@ -1,6 +1,6 @@
 import type * as Node from 'node:fs';
 import { pick } from 'utilium';
-import type { V_Context } from '../context.js';
+import type { V_Context } from '../internal/contexts.js';
 import type { InodeFields } from '../internal/inode.js';
 import { _inode_fields, hasAccess } from '../internal/inode.js';
 import * as c from './constants.js';
