+++ tests/fs/stat.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import { suite, test } from 'node:test';
-import { defaultContext } from '../../dist/internal/contexts.js';
+import { context } from '../../dist/context.js';
 import { Stats } from '../../dist/vfs/stats.js';
 import { fs } from '../common.js';
 
@@ -40,7 +40,7 @@ suite('Stats', () => {
 
 		fs.writeFileSync(newFile, 'hello', { mode: 0o640 });
 
-		const prevCredentials = { ...defaultContext.credentials };
+		const prevCredentials = { ...context.credentials };
 		const uid = 33;
 		const nonRootCredentials = {
 			uid,
@@ -53,7 +53,7 @@ suite('Stats', () => {
 
 		fs.chownSync(newFile, 0, nonRootCredentials.gid); // creating with root-user so that non-root user can access
 
-		Object.assign(defaultContext.credentials, nonRootCredentials);
+		Object.assign(context.credentials, nonRootCredentials);
 		const stat = fs.statSync(newFile);
 
 		assert.equal(stat.gid, nonRootCredentials.gid);
@@ -63,13 +63,13 @@ suite('Stats', () => {
 		assert.equal(stat.hasAccess(fs.constants.X_OK), false);
 		// changing group
 
-		Object.assign(defaultContext.credentials, { ...nonRootCredentials, gid: 44 });
+		Object.assign(context.credentials, { ...nonRootCredentials, gid: 44 });
 
 		assert.equal(stat.hasAccess(fs.constants.R_OK), false);
 		assert.equal(stat.hasAccess(fs.constants.W_OK), false);
 		assert.equal(stat.hasAccess(fs.constants.X_OK), false);
 
-		Object.assign(defaultContext.credentials, prevCredentials);
+		Object.assign(context.credentials, prevCredentials);
 	});
 
 	test('stat file', async () => {
